requirements.txt
sqlalchemy>=1.4
pandas>=1.3
requests>=2.25
python-dotenv>=0.21
openpyxl>=3.0  # for Excel export in pandas